requirements.txt
streamlit>=1.35.0
Pillow>=10.0.0
pytesseract>=0.3.10
easyocr>=1.7.1
requests>=2.31.0
python-dotenv>=1.0.0
ollama>=0.2.0
